Move StateBuilder to utils
