+++ tests/pypokerengine/api/emulator_test.py
@@ -1,12 +1,13 @@
 from nose.tools import raises
 from tests.base_unittest import BaseUnitTest
 from pypokerengine.api.emulator import Emulator, Event
-from pypokerengine.api.state_builder import restore_game_state, attach_hole_card,\
+from pypokerengine.utils.game_state_utils import restore_game_state, attach_hole_card,\
         attach_hole_card_from_deck, replace_community_card_from_deck
 from pypokerengine.engine.card import Card
 from pypokerengine.engine.pay_info import PayInfo
 from pypokerengine.engine.game_evaluator import GameEvaluator
 from pypokerengine.engine.poker_constants import PokerConstants as Const
+
 from examples.players.fold_man import PokerPlayer as FoldMan
 
 class EmulatorTest(BaseUnitTest):
